[Refactor] 빠른 검색 박스 컴포넌트화 완료
